backend_fastapi/requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-generativeai
email-validator
requests
